Update CreatesApplication.php
`getEnvironmentSetUp()` is a method where users can override to configure and setup configs. That is:
```php
/**
 * Define environment setup.
 *
 * @param  \Illuminate\Foundation\Application  $app
 * @return void
 */
protected function getEnvironmentSetUp($app)
{
    // Setup default database to use sqlite :memory:
    $app['config']->set('database.default', 'testbench');
    $app['config']->set('database.connections.testbench', [
        'driver'   => 'sqlite',
        'database' => ':memory:',
        'prefix'   => '',
    ]);
}
```

However, within the source code, I noticed it's called AFTER the service providers are bootstrapped.
The problem with that is if the Service Providers uses the config itself, then errors and issues would arise.